feature: Launcher application skeleton.
This is the skeleton of the launcher application.

Currently it makes basic tests. It makes the following assumptions:

http://dl-cdn.alpinelinux.org/alpine/edge/testing/ as been added
to the packages sources.

The following packages are already installed in the system:

```
apk add cri-o kubelet kubeadm kubelet-openrc cri-o-contrib-cni util-linux-misc
```

Ideally, the `kubeadm config images pull` has also been executed in order
to pre-download the cluster base images. In order to do that, CRI-O must be run.
Not sure this is possible in chroot.