✨ Option to wait for resources to settle
Currently it only waits for pods to be ready. It should be
smarter and also check the replicasets. As the provisioning
returns the provisionned resource ids, it should be possible
to check the deployments/statefulsets/daemonsets.
However, argo cd applications would also need to be taken
into account. Any kind of resource that can progress, in fact.