🐛 Initial configuration error (Admission Webhook)
MetalLB defines admission webhooks for custom resources.
One need to wait for the controller to be available before
creating the IPAddressPool and L2Advertisement resource.

In consequence, instead of creating the resources in Kustomize, we create
a Job that waits for the controller to be available before  creating the
resources.