✨ Can use fixed ip address and domain name
The ip address can be passed as the `--ip`  argument to the `start`
command.
If the address does not exist, it can be added (--ip-create) to the
interface pointed by `--ip-network-inteface` (defafult `eth0`).
The domain name of the cluster and node can be passed with `--domain-name`.

The configuration can be added to the `iknite.yaml` file:

```yaml
cluster:
  ip: 192.168.99.2
  create_ip: true
  network_interface: eth0
  domain_name: kaweezle.local
  kubernetes_version: 1.25.0
```

Now the configuration file can be in `~/.config/iknite/iknite.yaml` as well as
in `/etc/iknite/iknite.yaml`

TODO: adding the domain name to the machine if it doesn't exist.