✨ Proper flannel configuration
The previous flannel configuration was working but in case there was
other nodes in the cluster, it wouldn't work properly. Flannel works in
two parts:

- As a CNI plugin for assigning IP addresses to pods on a node.
- As a daemon for assigning a subnet to a node and make traffic flow
  from one node to the other.

What was working was the CNI configuration, but not the inter-node
routing. For this to work properly, the flannel daemon must access the
K8S API.

With this commit, the CNI part is still handled out of the cluster for
faster cluster initialization but the in-cluster daemon is added.

Also, the subnet.env file location is moved from `/run` in order to
avoid deletion issues.