✨ Ensure kubelet service is blocked and swallow existing kubelet
Internalize preventing the kubelet service from starting. The kubeadm
code starts the kubelet service but we want to handle that in our
start process. In order to prevent openrc to start the kubelet, we
add in `/etc/rc.conf` a non-exsiting dependency that makes the
start fail. Now some code ensures that the appropriate line is
present.

If iknite fails it will be restarted by openrc. If the previous process
has already started the kubelet, it will fail. Instead, we _swallow_ the
existing process so we can stop it when we are requested to stop.